+++ Soulworker Utility/UI/Option.cpp
@@ -7,7 +7,7 @@
 #include ".\Damage Meter\Damage Meter.h"
 #include ".\Buff Meter\Buff Meter.h"
 
-UiOption::UiOption()  :  _open(0), _framerate(1), _windowBorderSize(1), _fontScale(1), _columnFontScale(1), _tableFontScale(1), _is1K(0), _is1M(0), _isSoloMode(0), _hideName(0), _isTopMost(true), _cellPadding(0, 0), _windowWidth(800), _refreshTime(0.3) {
+UiOption::UiOption()  :  _open(0), _framerate(1), _windowBorderSize(1), _fontScale(1), _columnFontScale(1), _tableFontScale(1), _is1K(0), _is1M(0), _isSoloMode(0), _hideName(0), _isTopMost(true), _cellPadding(0, 0), _windowWidth(800), _refreshTime((FLOAT)0.3) {
 	_jobBasicColor[0] = ImVec4(ImGui::ColorConvertU32ToFloat4(ImColor(153, 153, 153, 255)));	// Unknown
 	_jobBasicColor[1] = ImVec4(ImGui::ColorConvertU32ToFloat4(ImColor(247, 142, 59, 255)));	// 하루
 	_jobBasicColor[2] = ImVec4(ImGui::ColorConvertU32ToFloat4(ImColor(59, 147, 247, 255)));	// 어윈
@@ -136,7 +136,7 @@ VOID UiOption::Helper() {
 	for (INT i = 0; i < 4; i++) {
 		sprintf_s(name, 128, "%s %d", STR_OPTION_TEST_VALUE_PLAYER, helper);
 		
-		UINT64 id;
+		UINT32 id;
 		if (helper == 3) {
 			id = DAMAGEMETER.GetMyID();
 		}
@@ -390,7 +390,7 @@ BOOL UiOption::GetOption() {
 		attr->QueryFloatValue(&winY);
 
 		//SetWindowPos(UIWINDOW.GetHWND(), HWND_NOTOPMOST, winX, winY, 0, 0, SWP_NOSIZE);
-		SetWindowPos(UIWINDOW.GetHWND(), HWND_TOPMOST, winX, winY, 0, 0, SWP_NOSIZE);
+		SetWindowPos(UIWINDOW.GetHWND(), HWND_TOPMOST, static_cast<INT>(winX), static_cast<INT>(winY), 0, 0, SWP_NOSIZE);
 
 #if DEBUG_READ_XML == 1
 		Log::WriteLog(const_cast<LPTSTR>(_T("Read WinPos(X,Y) = (%f, %f)")), winX, winY);
@@ -847,7 +847,7 @@ const FLOAT& UiOption::GetFramerate() {
 	return _framerate;
 }
 
-VOID UiOption::SetFramerate(UINT i) {
+VOID UiOption::SetFramerate(FLOAT i) {
 
 	if (i < 0)
 		i = 0;
